# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Indonesian Accounting System API
  description: |
    Professional accounting system API for Indonesian businesses.
    
    ## Authentication
    All endpoints except `/auth/login` and `/auth/register` require JWT authentication.
    Include the token in the Authorization header: `Bearer YOUR_TOKEN`
    
    ## Rate Limiting
    - Basic tier: 60 requests per minute
    - Premium tier: 300 requests per minute
    - API tier: 1000 requests per minute
    
    ## Indonesian Business Rules
    - All currency amounts are in Indonesian Rupiah (IDR)
    - Tax calculations follow Indonesian regulations (PPN 11%)
    - Date formats follow DD/MM/YYYY convention
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@accounting-system.co.id
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.accounting-system.co.id
    description: Production server
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: User login
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@contoh.co.id
                password:
                  type: string
                  format: password
                  example: SecurePassword123!
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      user:
                        $ref: '#/components/schemas/User'
                      company:
                        $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: admin@contoh.co.id
        name:
          type: string
          example: Administrator
        role:
          type: string
          enum: [admin, manager, accountant, user]
          example: admin
        company_id:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          nullable: true
    Company:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: PT Contoh Indonesia
        tax_id:
          type: string
          pattern: '^\d{2}\.\d{3}\.\d{3}\.\d{1}-\d{3}\.\d{3}$'
          example: 01.234.567.8-901.000
        address:
          type: string
          example: Jakarta, Indonesia
        phone:
          type: string
          example: +62-21-1234567
        email:
          type: string
          format: email
          example: admin@contoh.co.id
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: An error occurred
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: req_123456
        trace_id:
          type: string
          example: trace_789012
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until next request allowed
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when window resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'